% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{DataFrame$get_column}
\alias{DataFrame$get_column}
\title{Extract one or several columns}
\description{
Extract one or several columns
}
\examples{
DataFrame$new()$get_column("a")
DataFrame$new()$get_columns("a", "b")
}
