% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\docType{data}
\name{DataFrame-class}
\alias{DataFrame-class}
\alias{DataFrame}
\title{General class for a Polars DataFrame}
\format{
An object of class \code{environment} of length 4.
}
\usage{
DataFrame
}
\description{
General class for a Polars DataFrame
}
\keyword{datasets}
