% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame-double-brackets}
\alias{DataFrame-double-brackets}
\alias{$.DataFrame}
\alias{[[.DataFrame}
\title{Subset a Polars DataFrame}
\usage{
\method{$}{DataFrame}(self, name)

\method{[[}{DataFrame}(self, name)
}
\description{
Subset a Polars DataFrame
}
