% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame-dtypes}
\alias{DataFrame-dtypes}
\title{Get the type of variables in a Polars DataFrame}
\usage{
<DataFrame>$dtypes(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Get the type of variables in a Polars DataFrame
}
\examples{
DataFrame$new()$dtypes()
}
