% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame-get_column}
\alias{DataFrame-get_column}
\title{Extract columns from a Polars DataFrame}
\usage{
<DataFrame>$get_column(...)
<DataFrame>$get_columns(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Extract columns from a Polars DataFrame
}
\examples{
DataFrame$new()$get_column("a")
DataFrame$new()$get_columns("a", "b")
}
