% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\docType{data}
\name{pl}
\alias{pl}
\title{The complete polars public API.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pl
}
\description{
The complete polars public API.
}
\keyword{datasets}
