% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\docType{data}
\name{Series-class}
\alias{Series-class}
\alias{Series}
\title{Create new Series}
\format{
An object of class \code{environment} of length 3.
}
\usage{
Series(x, name)
}
\arguments{
\item{x}{any vector}

\item{name}{string}
}
\description{
Found in api as pl$Series named Series_constructor internally
}
\examples{
pl$Series(1:4)
}
\keyword{datasets}
